Abstracted both gen_ch_id_to_label and get_challenge_label_from_challenge
Some duplicated code which computes the challenge label from the challenge entry is now abstracted in its own method.

Also, community_prefix is now used instead of community acronym.

The computed mapping from challenge id to challenge label has been extracted from graphql_query_OEB_DB, so it is computed only when it is needed.

Both dataset and test action original ids are now checked with several rules which depend on the kind of entry and type.

Last, but not the least important, false positive mismatches due loss of precision are now detected and ignored.