Keys `data_model_repo` and `data_model_tag` are now optional.
These keys from submission_form_schema.json are used to tell
the repository where the JSON Schemas to be used for the
different validations of the generated scientific benchmarking
data model entries are. Now, when one or both are not declared,
the program fetches from the OpenEBench API the stored copy of
the schemas.

Also, added key `fix_original_ids` to submission_form_schema.json,
in preparation for the implementation of this new feature.